+++ frontend/websockets/old/index.html
@@ -1,0 +1,41 @@
+<html>
+
+<head>
+<meta charset="UTF-8">
+<meta name="description" content="This a University project for UPM">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<meta name=”keywords” content="IOTADB">
+<link rel="stylesheet" href='main.css'>
+</head>
+
+<body>
+    <div>
+        <input type='text' id='regName'>Name</input>
+        <input type='text' id='regPassword'>Password</input>
+        <input type='text' id='regEmail'>Email</input>
+        <button id='register' type='submit'>Register</button>
+    </div>
+    <div id='forReg'></div>
+    <div>
+        <input type='text' id='logName'>Name</input>
+        <input type='text' id='logPassword'>Password</input>
+        <button id='login' type='submit'>Login</button>
+    </div>
+    <div id='forLog'></div>
+
+    <div id='addresses_container'>
+    <input type='checkbox' id='addressesbox' value='true' onclick="addkey('addresses')">
+    <input type='text' id='addresses'></input></br></div>
+    
+    <div id='tags_container'>
+    <input type='checkbox' id='tagsbox' value='true' onclick="addkey('tags')">
+    <input type='text' id='tags'></input></br></div>
+   
+    <div><button id='submit' type='submit'>Search</button></div>
+    <div id='output_container'></div>
+
+    <script src='/socket.io/socket.io.js'></script>
+    <script src='index.js'></script>
+</body>
+
+</html>
